+++ copy-schema-files.js
@@ -11,7 +11,7 @@ const fs = require("fs");
         fs.promises.copyFile("demo/api/block-meta.json", "demo/site/block-meta.json"),
         fs.promises.copyFile("demo/api/schema.gql", "demo/admin/schema.gql"),
         fs.promises.copyFile("demo/api/schema.gql", "demo/site/schema.gql"),
-        fs.promises.copyFile("demo/api/comet-config.json", "demo/site/comet-config.json"),
-        fs.promises.copyFile("demo/api/comet-config.json", "demo/admin/comet-config.json"),
+        fs.promises.copyFile("demo/api/src/comet-config.json", "demo/site/src/comet-config.json"),
+        fs.promises.copyFile("demo/api/src/comet-config.json", "demo/admin/src/comet-config.json"),
     ]);
 })();
