+++ demo/admin/.gitignore
@@ -12,5 +12,5 @@ lang-compiled
 graphql.generated.ts
 block-meta.json
 src/blocks.generated.ts
-comet-config.json
+src/comet-config.json
 src/**/*.generated.ts
